pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi-agent-cli"
version = "1.0.0"
description = "Command-line tool for orchestrating multiple AI agents in parallel or sequence"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Kevin McAllorum", email = "kevin@mcallorum.com"},
]
keywords = [
    "ai",
    "agents",
    "cli",
    "orchestration",
    "pytest-agents",
    "automation",
    "devops",
    "multi-agent",
    "prometheus",
    "metrics",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pytest-agents>=1.0.0",
    "pyyaml>=6.0.0",
    "prometheus-client>=0.19.0",
    "typing-extensions>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
multi-agent-cli = "multi_agent_cli.cli:cli"

[project.urls]
Homepage = "https://github.com/kmcallorum/multi-agent-cli"
Documentation = "https://github.com/kmcallorum/multi-agent-cli#readme"
Repository = "https://github.com/kmcallorum/multi-agent-cli"
Issues = "https://github.com/kmcallorum/multi-agent-cli/issues"
Changelog = "https://github.com/kmcallorum/multi-agent-cli/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/multi_agent_cli"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
    "S",      # flake8-bandit (security)
]
ignore = [
    "S101",   # assert allowed in tests
    "TCH001", # type checking imports - not needed for this project
    "TCH002", # type checking imports - not needed for this project
    "TCH003", # type checking imports - not needed for this project
    "S104",   # binding to all interfaces is intentional for metrics
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG001", "ARG002", "S108"]
"src/multi_agent_cli/cli.py" = ["ARG001"]  # Click decorators require args

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "pytest_agents.*",
    "prometheus_client.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "requires_agents: Tests requiring real agents",
]
addopts = "-v --strict-markers"

[tool.coverage.run]
source = ["src/multi_agent_cli"]
branch = true
omit = []

[tool.coverage.report]
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "\\.\\.\\.",
]
show_missing = true
